Move script gpu files into webgpu folder
