+++ components/script/dom/mod.rs
@@ -327,46 +327,6 @@ pub mod gamepadevent;
 pub mod gamepadhapticactuator;
 pub mod gamepadpose;
 pub mod globalscope;
-pub mod gpu;
-pub mod gpuadapter;
-pub mod gpuadapterinfo;
-pub mod gpubindgroup;
-pub mod gpubindgrouplayout;
-pub mod gpubuffer;
-pub mod gpubufferusage;
-pub mod gpucanvascontext;
-pub mod gpucolorwrite;
-pub mod gpucommandbuffer;
-pub mod gpucommandencoder;
-pub mod gpucompilationinfo;
-pub mod gpucompilationmessage;
-pub mod gpucomputepassencoder;
-pub mod gpucomputepipeline;
-pub mod gpuconvert;
-pub mod gpudevice;
-pub mod gpudevicelostinfo;
-pub mod gpuerror;
-pub mod gpuinternalerror;
-pub mod gpumapmode;
-pub mod gpuoutofmemoryerror;
-pub mod gpupipelineerror;
-pub mod gpupipelinelayout;
-pub mod gpuqueryset;
-pub mod gpuqueue;
-pub mod gpurenderbundle;
-pub mod gpurenderbundleencoder;
-pub mod gpurenderpassencoder;
-pub mod gpurenderpipeline;
-pub mod gpusampler;
-pub mod gpushadermodule;
-pub mod gpushaderstage;
-pub mod gpusupportedfeatures;
-pub mod gpusupportedlimits;
-pub mod gputexture;
-pub mod gputextureusage;
-pub mod gputextureview;
-pub mod gpuuncapturederrorevent;
-pub mod gpuvalidationerror;
 pub mod hashchangeevent;
 pub mod headers;
 pub mod history;
@@ -619,6 +579,8 @@ pub mod websocket;
 mod webxr;
 #[cfg(feature = "webxr")]
 pub use self::webxr::*;
+mod webgpu;
+pub use self::webgpu::*;
 pub mod wheelevent;
 pub mod window;
 pub mod windowproxy;
