+++ components/script/dom/webgpu/gpuadapterinfo.rs
@@ -5,10 +5,9 @@
 use dom_struct::dom_struct;
 use webgpu::wgt::AdapterInfo;
 
-use super::bindings::codegen::Bindings::WebGPUBinding::GPUAdapterInfoMethods;
-use super::bindings::reflector::reflect_dom_object;
-use super::bindings::root::DomRoot;
-use crate::dom::bindings::reflector::Reflector;
+use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUAdapterInfoMethods;
+use crate::dom::bindings::reflector::{reflect_dom_object, Reflector};
+use crate::dom::bindings::root::DomRoot;
 use crate::dom::globalscope::GlobalScope;
 use crate::test::DOMString;
