+++ components/script/dom/webgpu/gpucanvascontext.rs
@@ -18,23 +18,20 @@ use webgpu::{
 use webrender_api::units::DeviceIntSize;
 use webrender_api::ImageKey;
 
-use super::bindings::codegen::Bindings::WebGPUBinding::{
-    GPUCanvasAlphaMode, GPUTextureFormat, GPUTextureUsageConstants,
-};
-use super::bindings::codegen::UnionTypes::HTMLCanvasElementOrOffscreenCanvas;
-use super::bindings::error::{Error, Fallible};
-use super::bindings::root::MutNullableDom;
-use super::bindings::str::USVString;
 use super::gpuconvert::convert_texture_descriptor;
 use super::gputexture::GPUTexture;
 use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUTexture_Binding::GPUTextureMethods;
 use crate::dom::bindings::codegen::Bindings::WebGPUBinding::{
-    GPUCanvasConfiguration, GPUCanvasContextMethods, GPUDeviceMethods, GPUExtent3D,
-    GPUExtent3DDict, GPUObjectDescriptorBase, GPUTextureDescriptor, GPUTextureDimension,
+    GPUCanvasAlphaMode, GPUCanvasConfiguration, GPUCanvasContextMethods, GPUDeviceMethods,
+    GPUExtent3D, GPUExtent3DDict, GPUObjectDescriptorBase, GPUTextureDescriptor,
+    GPUTextureDimension, GPUTextureFormat, GPUTextureUsageConstants,
 };
+use crate::dom::bindings::codegen::UnionTypes::HTMLCanvasElementOrOffscreenCanvas;
+use crate::dom::bindings::error::{Error, Fallible};
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::reflector::{reflect_dom_object, DomObject, Reflector};
-use crate::dom::bindings::root::{DomRoot, LayoutDom};
+use crate::dom::bindings::root::{DomRoot, LayoutDom, MutNullableDom};
+use crate::dom::bindings::str::USVString;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::htmlcanvaselement::{HTMLCanvasElement, LayoutCanvasRenderingContextHelpers};
 use crate::dom::node::{document_from_node, Node, NodeDamage};
