+++ components/script/dom/webgpu/gpuoutofmemoryerror.rs
@@ -5,12 +5,12 @@
 use dom_struct::dom_struct;
 use js::rust::HandleObject;
 
-use super::types::GPUError;
 use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUOutOfMemoryError_Binding::GPUOutOfMemoryErrorMethods;
 use crate::dom::bindings::reflector::reflect_dom_object_with_proto;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::globalscope::GlobalScope;
+use crate::dom::types::GPUError;
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
