+++ components/script/dom/webgpu/gpusupportedlimits.rs
@@ -7,10 +7,9 @@ use num_traits::bounds::UpperBounded;
 use webgpu::wgt::Limits;
 use GPUSupportedLimits_Binding::GPUSupportedLimitsMethods;
 
-use super::bindings::codegen::Bindings::WebGPUBinding::GPUSupportedLimits_Binding;
-use super::bindings::reflector::reflect_dom_object;
-use super::bindings::root::DomRoot;
-use crate::dom::bindings::reflector::Reflector;
+use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUSupportedLimits_Binding;
+use crate::dom::bindings::reflector::{reflect_dom_object, Reflector};
+use crate::dom::bindings::root::DomRoot;
 use crate::dom::globalscope::GlobalScope;
 
 #[dom_struct]
